op3: Hex-edit camera HAL to always disable QuadraCFA mode
QuadraCFA mode is randomly getting enabled, which causes our front-facing
camera to stop taking pictures. Hitting the shutter button while QuadraCFA
mode is enabled results in the camera stalling and no picture getting saved.

To fix this, hex-edit the camera HAL to make the setQuadraCfa() and
setQuadraCfaMode() subroutines do nothing. The changes are
as follows: at file offset 0xafe9e, the two bytes "50 B1" are replaced with
"C0 46". The "50 B1" corresponds the conditional branch in setQuadraCfa()
that avoids the "Quadra CFA mode not supported" error condition. By replacing
that with "C0 46" (no-op instruction), the non-fatal
"Quadra CFA mode not supported" error condition is always triggered.

The next change is replacing the two bytes "49 D0" at file offset 0xafaf8 with
"82 E0". The "49 D0" corresponds to the conditional branch in setQuadraCfaMode()
that's _supposed_ to cause the entire subroutine to be skipped when getQuadraCfa()
returns 0 (false). Although the hex-edit above made to setQuadraCfa() makes
getQuadraCfa() always return 0, OnePlus modified the setQuadraCfaMode() subroutine
so that not all of it is skipped when getQuadraCfa() returns 0, so this additional
change is needed just to be safe. The "82 E0" corresponds to an unconditional
branch that skips the entire setQuadraCfaMode() subroutine.